vinext migration: MDX parity via explicit @mdx-js/rollup (#2)
Re-establish the remark/rehype plugin chain under vinext (Vite). Until
this commit, vinext's MDX support was running but with no custom
plugins — its build log showed `[vinext] Auto-injected @mdx-js/rollup
for on-demand MDX support`, i.e. the bare default factory with no
options. The `@next/mdx` webpack-probe path that vinext uses to lift
remark/rehype config off next.config never fired the
"...with remark/rehype plugins from next.config" branch, so the custom
TOC, slug, autolink-headings, and highlight transforms were silently
absent.

Key decisions:
- Stop relying on the webpack-probe path. Register `@mdx-js/rollup`
  explicitly in vite.config.ts. vinext detects a user-supplied MDX
  plugin via `hasUserMdxPlugin` (`p.name === "@mdx-js/rollup" || "mdx"`)
  and short-circuits its own auto-injection, so there's no double-
  compile. Confirmed: the on-demand log line is gone from
  `vinext build` output; the build shows `- @mdx-js/rollup (5%)` in the
  transform progress.
- Use `enforce: "pre"` on the @mdx-js/rollup plugin to match vinext's
  own `vinext:mdx` ordering, so .mdx transforms happen before other
  plugins that expect JS/JSX input.
- Extract the plugin chain (createRemarkToc + remarkFrontmatter +
  rehypeSlug + custom rehypeAutolinkHeadings + rehypeHighlight) into
  shared `mdx.config.mjs`. Both next.config.mjs (legacy `next build`
  path, kept until slice #8 cleanup) and vite.config.ts import from
  the same source. One chain, two consumers — no drift between them.
- Cast the imported plugin arrays at the vite.config.ts boundary
  because mdx-js's `@types/unified` (vfile v6) clashes with the
  inferred unified types pulled in from peer deps (vfile v5). Runtime
  shapes are correct; only TS type identity diverges.
- next.config.mjs slimmed: dropped the inline plugin imports and the
  `createRemarkToc` definition (now in mdx.config.mjs). Behavior
  unchanged for the legacy webpack path.

Verified output parity by grepping prerendered RSC server chunks:
- `header-anchor` class and `header-anchor-wrapper-{h2,h3}` group
  classes are baked in -> rehype-autolink-headings ran with the
  custom `behavior:'after'`, `properties.class`, `content`, `group`,
  `test:['h2','h3']` config.
- A page using `## Table of Contents` renders as
  `<h2>Table of Contents</h2><ul><li><a href="#overview">…</a></li>…`
  -> createRemarkToc walked the mdast and injected the TOC map.
- No raw `meta:`/`title:` frontmatter strings leak into rendered
  output -> remark-frontmatter strips the YAML block.
- pageExtensions still includes md/mdx (kept in next.config.mjs).

Force-static routes verified at build classification:
- /resources/:slug*, /:slug+ (simple-mdx), /podcast, /podcast/:slug,
  /newsletter, /newsletter/issues/:slug, /members all classify as
  ○ Static. vinext prerenders Static routes at build time in Node, so
  the fs.readdirSync/readFileSync/process.cwd() calls in
  src/util/loadMdx.server.ts and the `path` import in
  src/components/PostList.tsx run during the build, not on the
  Workers runtime. The path-externalized-for-browser warning on
  PostList.tsx during the client build is benign — that module is
  only imported from server-rendered pages.

Files changed:
- mdx.config.mjs: new — single source of truth for remark/rehype
  plugin chain, including createRemarkToc
- next.config.mjs: import plugins from mdx.config.mjs; drop inline
  plugin imports and createRemarkToc body
- vite.config.ts: add @mdx-js/rollup as an explicit `enforce:"pre"`
  plugin wired with the shared chain; type-cast plugin arrays to
  bridge the vfile-version skew

Verified:
- pnpm typecheck clean (after removing the stale Next 15 .next/types/
  validator.ts file, same situation as noted in the scaffold commit;
  .next is gitignored)
- pnpm build:vinext completes; "on-demand MDX support" log line gone;
  routes classified as in earlier slices
- No `pnpm test` script defined in package.json — nothing to run

Blockers / notes for next iteration:
- pnpm lint remains broken on main from the eslint-config-next +
  react 19 circular-config bug, unrelated to this change.
- vinext start currently errors locally with
  `ERR_UNSUPPORTED_ESM_URL_SCHEME` on `cloudflare:` — expected, since
  Node's default ESM loader can't resolve `cloudflare:workers` outside
  the Workers runtime. Real serve happens via wrangler/Workers in
  slice #8; not in scope here.
- Side-by-side screenshot diffs vs current prod (resources/handbook/
  newsletter) are deferred to the operator at preview-deploy time —
  the static-analysis evidence above covers the per-plugin output
  parity, but human eyes are still the right check before cutover.
